+++ src/main/java/gift/DTO/ResponseOptionDTO.java
@@ -1,4 +1,4 @@
-package gift.Model;
+package gift.DTO;
 
 public class ResponseOptionDTO {
     private Long id;
